get_all_devices ignores nic if mac not present
The get_all_devices method of the dpm agent returns a list
of mac addresses that belong to NICs it manages. As the
mac address is not a DPM NIC attribute, Nova stores it
in the NICs description field. If that mac was not present
dpm agent added "None" to the device_list, which caused
a couple of errors in the logs.

Change-Id: If12bbbaf8cc7209d0554956ac7faba4d8845e0f0
Closes-Bug: #1655675