Using zhmcclient via global-requirements
zhmcclient is now an offical OpenStack package managed
by OpenStack requirements [1]. Let's use it!

To allow newer versions in the future, an update to the
upper-constraints.txt in the OpenStack requirements
repo is required.

To bump the minimum just increase the version in requirements.txt
of networking-dpm and make sure, that newer versions are allowed
by OpenStack requirements (see above).

[1] https://review.openstack.org/413689

Change-Id: I931b0e1791a3a4cb54328e8489419ebb7fc385dc
Closes-Bug: #1656244