Revert "Include common agent configs in config file gen"
This reverts commit 9a347ef9c8535c2bd51e401056153a0c50a9b1f5.

This commit introduced a dependency to Neutron in the configuration
definition file (config.py). Due to that, the RTD build is failing,
as Neutron is not being installed (as not listed in
test-requirements.txt)

This revert removes that dependency again. As a result, only the dpm
config options can be seen in autogenerated doc, but not the
also interessting neutron options. Those need to be described
manually now.

Change-Id: I106993a07f66d877cd6ae397c7011dfa719b9a2b
Partial-Bug: #1660572