Remove VLAN support
This patch removes the VLAN support from the DPM Neutron
mechanism driver.

DPM adapters do not support VLAN tagging. Therefore the VLAN tagging
must be done from inside the operatin system via a VLAN device.
For automated configuration the operating system must be aware of the
VLAN ID to be used.

1) Ideally this information is part of the config drive. Then
cloud-init can do all the configuration automatically.

2) Alternatively the vlan-id could be passed in via the
boot-os-specific-parameters property of a partition. But then some
code inside of the operating system is required to interpret this
data and configure the VLAN device. It's questionable if cloud-init
would be able to deal with that.

Due to 2) and to the fact, that Ocata release is near, we agreed
to remove the VLAN support again. If one tries to attache an instance
to a VLAN network, it will result in an error message. This is a
clear statement that this is not supported.

The VLAN code should be added again during Pike. Ideally in
combination with the config drive 1). In the worst case using
approach 2).

Closes-Bug: #1661239
Change-Id: I0b4d86fc98eda0b67b084903d0d3ce1ba9f16cd8