Continue agent when HMC connection is lost
If the agent lost its connection to the HMC (for whatever reason)
it terminated. This patch allows the agent to continue working.
In this situation the agent reports all NICs of the current vswitch
as DOWN, which also causes the corresponding Neutron ports to be
set to DOWN state. If the HMC is not reachable for a longer time
this will result in all NICs to be reported as DOWN.

Once the HMC revives, the NICs and corresponding
ports willl be reported as UP again.

Change-Id: Ie6753c13f6d53df1e58c534ebb3eb36a7cab4140
Closes-Bug: #1655866