Continue agent when HMC connection is lost
If the agent loses its connection to the HMC (for whatever reason)
it terminated. This patch allows the agent to continue working in
this situation. In such a situation the agent reports all NICs
as DOWN, which also causes the corresponding Neutron ports to be
set to DOWN state. Once the HMC revives, the NICs and corresponding
ports willl be reported as UP again.

Change-Id: Ie6753c13f6d53df1e58c534ebb3eb36a7cab4140
Closes-Bug: #1655866