Continue agent when HMC connection is lost
If the agent lost its connection to the HMC (for whatever reason)
it terminated. This patch allows the agent to continue working.
In this situation the agent reports all devices (NICs)
as DOWN, which also causes the corresponding Neutron ports to be
set to DOWN state.

Once the HMC revives, the NICs and corresponding
ports willl be reported as UP again.

This is a temporary solution. In the long run having something like a
'retry' and/or a 'timeout' makes sense.

Change-Id: Ie6753c13f6d53df1e58c534ebb3eb36a7cab4140
Closes-Bug: #1655866
(cherry picked from commit 3357e13a4d31782dfb6d7f431b2fa4e0b56a9edb)