Fix "No such file error" on stack.sh
When doing a stack.sh, the following error occured:

    tee: //etc/neutron/plugins/ml2/dpm_agent.ini:
     No such file or directory

This happened as the devstack plugin tried to write the
firewall_driver config option, although the target file has
not yet been created. As this option is obsolete (if non is
set, the noop driver is used - that's what we wanted) we can
just get rid of it.

Change-Id: I0a52f3dccd8fc44cfb65d1548dfb529b268bcd82
Closes-Bug: #1665600