Treat port-element id as String
The HMC and requires port-element-id to be a string!

The port-element-id taken from the config file got converted from
string to int. The code internally used the int value. But for
communication to the HMC it was converted to string again.

This patch removes the transiation from string - int - string. Now
always the string value is used.

Change-Id: I51361f6abd2b0b447ecbf0844a3a38459999678f