Introduce oslo config object for adapter_mapping
This patch introduces a oslo config object for a network adapter port
mapping. It encapsulates the validation of the config value and does
the parsing. It's a multi value config option.

It requires the following arguments

* name: The name of the config option.

The config value must be of the following format in the config file

  <name>=<physnet>:<adapter-id>:<port>
  <name>=>physnet>:<adapter-id>:
  <name>=>physnet>:<adapter-id>
  # (multi valued)

It returns a list of tuples:

  [(<physnet>, <adapter-id>, <port>),
   (<physnet>, <adapter-id>, <port>),
   (<physnet>, <adapter-id>, <port>)]

where

* physnet:    The physical network string. Can not be 'None'.
              No restriction in length. Keeps case from config file.
* adapter-id: The object-id of the adapter that should be used.
              Cannot be 'None'. Will be returned in lower case.
* port:       The port-element-id string to be used. Defaults to '0'.
              Cannot be 'None'

Change-Id: I407bda2896d667ef6fad4138a9787da4b39a19f7
Partial-Bug: #1663369
Partial-Bug: #1659315