Cleanup requirements files
networking-dpm requires Neutron and its requirements to be present.
Neutron gets installed via the 'tox_install.sh' script when tox
is exectuted. Networking-dpms requirements must always be compatible
with the requirements of Neutron. At the moment many requirements
are listed in networking-dpms and Neutrons requirements. And both
must always be in sync (no contradictions).

Problems occur if the version of a requirement is changed. The new
version cannot safely be merged in networking-dpm until Neutron
picked it up.

To avoid this, this patch cleans up all duplicate requirements. Now
networking-dpm only lists requirements that are required on top
of Neutron. New versions of common libraries get applied when
they have merged in Neutron.

Change-Id: I46d6e8b7fa83a8eaa1fd1ba4b5becda8dd87dfaa
Closes-Bug: #1669695