[WIP] Cleanup requirements files
Networking-dpm requires Neutron and its requirements to be present.
At the moment many requirements are listed in networking-dpms and
Neutrons requirements files (requirements.txt and
test-requirements.txt). For a clean and working project setup, both
files must always be in sync (no contradictions!).

Problems occur if the version of a requirement is changed. The new
version cannot safely be merged in networking-dpm until Neutron
picked it up. To avoid this, this patch cleans up all duplicate
requirements. Now networking-dpm only lists requirements that are
required on top of Neutron. New versions of common libraries get
applied when they have merged in Neutron.

In the tox environment, requirements in Neutron requirements.txt
already get installed by the *install_tox.sh* script on every tox
environment build. This is not the case for the test-requirements.
This patch also installs the test requirements of Neutron on
every tox environment build.

Some requirements of the test-requirements file just were required
for the Read the docs (RTD) docbuild. This patch moves those into
a separate requirements file. This file can then be specified in
the RTD projects configuration (website). This new file is *not* used
in the OpenStack gate or for local testing using tox.

TBD: Will this file be updated by the Proposal Bot?

Change-Id: I46d6e8b7fa83a8eaa1fd1ba4b5becda8dd87dfaa
Closes-Bug: #1669695