fakezhmc: Enforce type for 'port' and 'element-id'
The HMC treates the 'port' attribute of a vswitch as int, but an
adapter ports 'element-id' as str. Both have the same value set.
This is confusing and might lead to problems if not considered
like it happened with bug [1].

This patch makes the fakezhmcclient aware of the different types and
enforces it. All tests that were still using wrong types also needed
aaption.

[1] https://bugs.launchpad.net/networking-dpm/+bug/1673185

Change-Id: Iede6e26a6ef94cac0587ab1ad147409bbb141778