Convert port to int for vswitch object
Patch [1] changed the behavior of 'port' to be always of type str.
Till then it was read as str from the config file, converted to int
and converted back to str for some cases.

But this patch did not consider that 2 flavors of port exists in
the DPM API:

* the 'port' attribute of a vswitch is of type int
* the 'element-id' of an adapter port is of type str

With [1], also for the vswitch case the int string was used, which
resulted in a zhmcclient.NotFound exception.

This patch fixes this by converting the value for this special
case to int.

Closes-Bug: #1673185

[1] https://github.com/openstack/networking-dpm/commit/f59ff63da45da03c3d01bc0fc561347609c97524

Change-Id: I6fbfbb8d81b5078ac336aab33da4295a9b835d9c