Ensure vif_details are unique per VIF
The Neutron framework assumes that the details of a vif are the same
cross all nics [1]. But with networking DPM they aren't! The vlan
detail might be different for each vif, depending to which network
it belongs.

The current implementation had a big bug: If first a vif attached to
a vlan network was processed, the vlan id is stored in the instance
drivers instance variable self.vif_details. Now if after that
a flat vif is processed, self.vif_details did still contain the
vlan property.

The solution is to make a copy of self.vif_details each time we process
it.

[1] https://github.com/openstack/neutron/blob/7e3d6a18fb928bcd303a44c1736d0d6ca9c7f0ab/neutron/plugins/ml2/drivers/mech_agent.py#L142,L143

Change-Id: I0cc09c20c308879e198a7fefd8c2d68c8591847e