Get tests going again
* Import "topics" from neutron_lib
* Import Neutron via requirements.txt

"topics" has been moved to neutron-lib recently [1]. Need to
import it from there.

The way of importing Neutron into a project has changed. Now as it is
released on pypi, we can use it from there. The old apporach,
specifying the requirement via a git url did not work anymore, since
pip install tried to apply the constraint of [2] on it.

[1] https://github.com/openstack/neutron/commit/ef93f7e7f0d74af36767b4dc1347b50a72a04463
[2] https://review.openstack.org/#/c/550475/

Change-Id: I62044f0878d1ebf12b73aacbd18996159878f797